Former-commit-id: 6adbb3f7026bf9d46ea9781842fab9c3b46c6416

+++ Sources/Internal/Base/FastName.cpp
@@ -110,7 +110,7 @@ void FastName::Init(const char * name)
         else
         {
             // index will be a new row in names table
-            index = db->namesTable.size();
+            index = static_cast<int32>(db->namesTable.size());
             db->namesTable.resize(index + 1);
             db->namesRefCounts.resize(index + 1);
         }
